auto-sort-notes/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pdfminer
pdfminer.six
pydantic
tqdm
python-magic; platform_system != "Windows"
python-magic-bin; platform_system == "Windows"